% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusing_adder.R
\name{confusing_adder}
\alias{confusing_adder}
\title{Adds seven to a variable in an intentionally confusing way}
\usage{
confusing_adder(x)
}
\arguments{
\item{x}{is an unquoted variable name}
}
\value{
the value of x with added 7
}
\description{
Adds seven to a variable in an intentionally confusing way
}
\examples{
x <- 10
confusing_adder(x)

y <- 20
confusing_adder(y)
}

