% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumfunction.R
\name{sumfunction}
\alias{sumfunction}
\title{Add eight to a number}
\usage{
sumfunction(x)
}
\arguments{
\item{x}{A number}
}
\value{
the sum of x and eight
}
\description{
Add eight to a number
}
\examples{
sumfunction(x)
}

